% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCgene.R
\name{FCgene}
\alias{FCgene}
\title{Fold Change Analysis for Gene Expression Profiling
Conduct the differential expression analysis using fold change method and find the (significantly) differential expressed genes(DEGs)}
\usage{
FCgene(exprSet, sig.level = 2, col.g1, col.g2, writeout = FALSE,
  q = 0.2, log.tran = FALSE)
}
\arguments{
\item{exprSet}{The gene expression profiling with each row represents a gene, each column represents an sample}

\item{sig.level}{The threshold for screening the significantly differential expressed gene, 2 (by default) is the typical value used in fold change methods.}

\item{col.g1}{A numeric vector indicates the columns that contain samples from the first group}

\item{col.g2}{A numeric vector indicates the columns that contain samples from the second group}

\item{writeout}{A logical value indicating whether a data frame contains name of the DEGS and their fold change result should output as a .csv file in to the working directory}

\item{q}{A number less than 1 indicate the quantile you want to set for clustering the fold change result several levels}

\item{log.tran}{A logical value indicating whether the data needs log transformation, which is what usually done in gene expression analysis.}
}
\value{
A list that contains a factor that contains the fold change results of all the , name of the DEGS, their fold change result
}
\description{
Fold Change Analysis for Gene Expression Profiling
Conduct the differential expression analysis using fold change method and find the (significantly) differential expressed genes(DEGs)
}
