% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCplot.R
\name{FCplot}
\alias{FCplot}
\title{Visualization of FCgene Object
Visualizing of the fold change result using a ggplot style pie chart}
\usage{
FCplot(x)
}
\arguments{
\item{x}{An output object from FCgene function in the FCgene package}
}
\description{
Visualization of FCgene Object
Visualizing of the fold change result using a ggplot style pie chart
}
\examples{
data(mydata)
MyexprSet <- mydata
MyFCgene <- FCgene(MyexprSet, col.g1 = c(1:4), col.g2 = c(5:8))
FCplot(MyFCgene)
}
