% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneprof-data.R
\docType{data}
\name{mydata}
\alias{mydata}
\title{Gene expression profiling example}
\format{A gene expression profiling}
\usage{
mydata
}
\description{
Gene expression profiling data from 8 ischemic stroke patients
}
\keyword{datasets}
